backend/requirements.txt
# Core framework
fastapi==0.116.1
uvicorn==0.35.0  # Used for running FastAPI apps

# ORM and database
sqlalchemy==2.0.41
alembic==1.16.4  # For database migrations

# Async Postgres driver
asyncpg==0.30.0

# Auth and security
authlib==1.6.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
cryptography==45.0.5

# Env and config management
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1

# File handling
aiofiles==24.1.0
python-multipart==0.0.20

# PDF processing
pdfminer.six==20250506

# HTTP and utility
httpx==0.28.1
markdown==3.8.2

# AI/NLP features
chromadb==1.0.15
sentence-transformers==5.0.0

# Starlette is a FastAPI dependency but used directly in your code too
starlette==0.46.2
